very long road for this one.
Such a long time solving this one. I needed to more to a more efficient
prime detector. Researching this led to the miller rabin prime detection
method. Coding it up was a bit of a challenge. After that things were
very fast.

I want to circle back to get a deeper understanding of how it works.
http://en.wikipedia.org/wiki/Miller%E2%80%93Rabin_primality_test